Implements Spire (#5)
TL;DR
-----

Provides an initial implementation for Spire based on FRSCA

Details
-------

Deploys Spire using the same Helm chart as FRSCA. The chart is vendored in using
the Helm tarball to address a mismatch between vendir's Helm vendoring and the
Replicated Helm implementation: when vendir pulls in a chart it unpacks it to
the source template, while Replicted expects a tarball.

The primary diference from the upstream FRSCA implementation is that the Spire
services is implmented as type LoadBalancer. I'd wanted to use ClusterIP
and use the ingress capabilities of the Helm chart, but ther are errors in the
Ingress template.